fix: clippy warnings and add mandatory CI hooks
CI/CD:
- Pre-push hook runs: format check, cargo check, clippy, tests
- Clippy allows patterns intentional in codebase
- Hook script in .githooks/pre-push with install.sh

Fixes:
- sr25519 migration for data_gossip.rs and auth.rs
- Dead code annotations for unused fields
- Consistent digit grouping in numbers
- Redundant field names removed
- Module naming conflict fixed
- Simplified JSON-RPC status code
- Updated Docker image whitelist

+++ crates/bittensor-integration/src/validator_sync.rs
@@ -125,9 +125,9 @@ impl ValidatorSync {
             }
 
             // Extract normalized scores (u16 -> f64, divide by u16::MAX)
-            let incentive = neuron.incentive as f64 / u16::MAX as f64;
-            let trust = neuron.trust as f64 / u16::MAX as f64;
-            let consensus = neuron.consensus as f64 / u16::MAX as f64;
+            let incentive = neuron.incentive / u16::MAX as f64;
+            let trust = neuron.trust / u16::MAX as f64;
+            let consensus = neuron.consensus / u16::MAX as f64;
 
             // Check if active (has stake)
             let active = stake > 0;
